[common] move new plotting setup to new file
